+++ migration/jobService.go
@@ -10,6 +10,7 @@ import (
 	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/fields"
 	watchAPI "k8s.io/apimachinery/pkg/watch"
 	"k8s.io/client-go/tools/watch"
 	"log/slog"
@@ -24,7 +25,7 @@ type getStreamerFunc func(jobName string, options *corev1.PodLogOptions) (stream
 
 // getWatcherFunc is a function type that creates a watcher for monitoring job status changes
 // It takes a context and resource version and returns a watch interface for receiving events
-type getWatcherFunc func(ctx context.Context, resourceVersion string) (watchAPI.Interface, error)
+type getWatcherFunc func(ctx context.Context, jobName string) (watchAPI.Interface, error)
 
 // JobServiceDependencies contains all the dependencies required to create a JobService
 // It includes dependencies for job creation, job client for interacting with Kubernetes jobs,
@@ -80,15 +81,8 @@ func NewJobService(deps JobServiceDependencies) (*JobService, error) {
 // 3. Returns a log streamer for that pod
 func createGetStreamerFunc(podClient podClient) getStreamerFunc {
 	return func(jobName string, options *corev1.PodLogOptions) (streamer, error) {
-		// Create a label selector to find pods belonging to the specified job
-		jobLabelSelector := &metav1.LabelSelector{
-			MatchLabels: map[string]string{
-				batchv1.JobNameLabel: jobName,
-			},
-		}
-
 		// List all pods with the job label
-		pods, err := podClient.List(context.Background(), metav1.ListOptions{LabelSelector: metav1.FormatLabelSelector(jobLabelSelector)})
+		pods, err := podClient.List(context.Background(), metav1.ListOptions{LabelSelector: buildJobLabelSelector(jobName)})
 		if err != nil {
 			return nil, fmt.Errorf("failed to list pods for job %s: %w", jobName, err)
 		}
@@ -114,15 +108,41 @@ func createGetStreamerFunc(podClient podClient) getStreamerFunc {
 // It returns a getWatcherFunc that creates a RetryWatcher, which automatically reconnects
 // if the watch connection is lost
 func createGetWatcherFunc(jobClient jobClient) getWatcherFunc {
-	return func(ctx context.Context, resourceVersion string) (watchAPI.Interface, error) {
+	return func(ctx context.Context, jobName string) (watchAPI.Interface, error) {
+		nameSelector := fields.OneTermEqualSelector("metadata.name", jobName).String()
+		jobList, err := jobClient.List(ctx, metav1.ListOptions{
+			FieldSelector: nameSelector,
+		})
+		if err != nil {
+			return nil, fmt.Errorf("error getting migration job while trying to create job watcher: %v", err)
+		}
+		if len(jobList.Items) == 0 {
+			return nil, fmt.Errorf("migration job %s not found while trying to create job watcher", jobName)
+		}
+		rv := jobList.ResourceVersion
+
+		slog.Info(fmt.Sprintf("get watcher func job name: %s, job list resource version: %s", jobName, rv))
 		// Create an adapter to make jobClient.Watch compatible with RetryWatcher
 		wrapper := watchAdapter{
-			watchFunc: jobClient.Watch,
+			watchFunc: func(ctx context.Context, opts metav1.ListOptions) (watchAPI.Interface, error) {
+				opts.FieldSelector = nameSelector
+				return jobClient.Watch(ctx, opts)
+			},
 		}
 
 		// Create a RetryWatcher that will automatically reconnect if the watch connection is lost
-		return watch.NewRetryWatcherWithContext(ctx, resourceVersion, wrapper)
+		return watch.NewRetryWatcherWithContext(ctx, rv, wrapper)
 	}
+}
+
+// Create a label selector to find resources belonging to the specified job
+func buildJobLabelSelector(jobName string) string {
+	jobLabelSelector := &metav1.LabelSelector{
+		MatchLabels: map[string]string{
+			batchv1.JobNameLabel: jobName,
+		},
+	}
+	return metav1.FormatLabelSelector(jobLabelSelector)
 }
 
 // Run creates and executes a Kubernetes job, watches for its completion, and returns its logs
@@ -167,7 +187,7 @@ func (j JobService) Run(ctx context.Context) (jobLogs io.ReadCloser, err error) 
 	}()
 
 	// Create a watcher to monitor the job's status
-	watcher, err := j.getWatcher(ctx, jobResource.GetResourceVersion())
+	watcher, err := j.getWatcher(ctx, jobResource.GetName())
 	if err != nil {
 		return nil, fmt.Errorf("failed to create watcher for job %s: %w", jobResource.GetName(), err)
 	}
@@ -180,7 +200,7 @@ func (j JobService) Run(ctx context.Context) (jobLogs io.ReadCloser, err error) 
 	slog.Debug("Starting to wait for job to complete or fail")
 
 	// Process events from the watcher until the job completes, fails, or an error occurs
-	errWatch := watchEvents(watcher.ResultChan())
+	errWatch := watchEvents(watcher.ResultChan(), jobResource.GetName())
 	if errWatch != nil {
 		return nil, fmt.Errorf("received error while watching job: %w", errWatch)
 	}
@@ -191,7 +211,7 @@ func (j JobService) Run(ctx context.Context) (jobLogs io.ReadCloser, err error) 
 
 // watchEvents processes events from a watcher until the job completes, fails, or an error occurs.
 // It logs the event type and returns an error if the job fails, or an error occurs during processing
-func watchEvents(resultChan <-chan watchAPI.Event) (errWatch error) {
+func watchEvents(resultChan <-chan watchAPI.Event, jobName string) (errWatch error) {
 	for event := range resultChan {
 		slog.Debug("Received event from watcher for import job", "type", event.Type)
 
@@ -205,6 +225,12 @@ func watchEvents(resultChan <-chan watchAPI.Event) (errWatch error) {
 		jobChange, ok := event.Object.(*batchv1.Job)
 		if !ok {
 			errWatch = fmt.Errorf("received unexpected event type during watch of job: %T", event.Object)
+			break
+		}
+
+		// ignore events from other jobs
+		// this is just a failsafe and should never happen
+		if !(jobChange.GetName() == jobName) {
 			break
 		}